+++ js/bingo/mv/command/event.js
@@ -2,8 +2,9 @@
 (function (bingo) {
     /*
         使用方法:
-        bg-event="{click:function(e){}, dblclick:helper.dblclick}"
+        bg-event="{click:function(e){}, dblclick:helper.dblclick, change:['input', helper.dblclick]}"
         bg-click="helper.click"     //绑定到方法
+        bg-click="['input', helper.click]"     //绑定到数组, 等效于$().on('click', 'input', helper.click)
         bg-click="helper.click()"   //直接执行方法
     */
     bingo.each('event,click,blur,change,dblclick,focus,focusin,focusout,keydown,keypress,keyup,mousedown,mouseenter,mouseleave,mousemove,mouseout,mouseover,mouseup,resize,scroll,select,submit,contextmenu'.split(','), function (eventName) {
@@ -15,17 +16,21 @@
                 /// <param name="$attr" value="bingo.view.viewnodeAttrClass()"></param>
 
                 var bind = function (evName, callback) {
-                    $node.on(evName, function () {
-                        //console.log(eventName);
-                        var r = callback.apply(this, arguments);
-                        $view.$update();
-                        return r;
-                    });
+                    if (bingo.isArray(callback))
+                        $node.on.apply($node, [].concat(evName, callback));
+                    else {
+                        $node.on(evName, function () {
+                            //console.log(eventName);
+                            var r = callback.apply(this, arguments);
+                            $view.$update();
+                            return r;
+                        });
+                    }
                 };
 
                 if (eventName != 'event') {
-                    var fn = $attr.$value();
-                    if (!bingo.isFunction(fn))
+                    var fn = /^\s*\[(.|\n)*\]\s*$/g.test($attr.$attrValue()) ? $attr.$results() : $attr.$value();
+                    if (!bingo.isFunction(fn) && !bingo.isArray(fn))
                         fn = function (e) { return $attr.$eval(e); };
                     bind(eventName, fn);
                 } else {
@@ -35,7 +40,7 @@
                         for (var n in evObj) {
                             if (bingo.hasOwnProp(evObj, n)) {
                                 fn = evObj[n];
-                                if (bingo.isFunction(fn))
+                                if (bingo.isFunction(fn) || bingo.isArray(fn))
                                     bind(n, fn);
                             }
                         }